The Usage of Valley Map data for surface T and Q observations in GSI for HRRR-based 3DRTMA (#854)
To improve the analysis of surface observations in complicated terrain
area, esp. in the valley regions, (2D)RTMA/URMA has been adopting a
"valley-map" technique (developed by Manuel Pondeca of NCEP/EMC) to
check the valley-like terrain at the observation location and relax the
observation gross error if the observation is in a valley-like area.
This "valley-map" technique is useful to more efficiently use the
surface observations near valley area in 2DRTMA/URMA, so it is decided
to add this capability in GSI for HRRR-based 3DRTMA.