Height Adjustment of Background 10-m Wind Gust When Analyzing Near-surface Wind Gust observations in GSI for HRRR-3DRTMA (#863)
- The height adjustment scheme is based on the Similarity Theory for
turbulence boundary layer. In fact, it already had been implemented in
GSI (by @ManuelPondeca-NOAA, @StevenLevine-NOAA and @xyzemc a few years
ago) for 2DRTMA/URMA (see the if-block of "if(zob < ten)" in
setupgust.f90). In this work, the same code would be also used for
3DRTMA run.
- Need to add code in GSI read in the Roughness Length (ZNT) from the
firstguess file, or using the default value of roughness set in GSI for
regional analysis (which is 0.05, might be too small for some area).