Add capability to run AutoQC package for 3D-RTMA/URMA (#883)
The changes are made to enable the inclusion of an automated quality
control package in 3D-RTMA/URMA. This involves reading and applying the
accept lists generated by the AutoQC package; a new option
(i_gsdsfc_uselist=2) was added to read these lists. In addition, a
mesonet wind bias correction algorithm is performed as part of the
AutoQC package; the GSI code modifications use the output of this
algorithm to adjust the raw wind speed & gust observations.