Comment out prepBUFR blacklist station prints + update fix hash (#919)
Closes #918 

This PR comments out the print in `read_prepbufr` that explicitly prints
out each station observation that is found that is to be rejected by the
prescribed blacklist. This is commented out rather than removed so that
it can be enabled later for debugging purposes. This PR also updates the
fix submodule to make files consistent with or without 2m/soil DA.