GSI for 3DRTMA: Enhancement of Valley-Check for Surface Obs of T & Q and Applying GSD Terrain Match to MESONET Obs of T (kx=188)  (#957)
1. The valley-map adjusted usage value of surface observation is very
useful information in the ObsQC of RTMA. Recently RTMA colleague
reported that for some surface observations of Q, their valley-adjusted
usage values were lost and reset to some other values. Therefore the
efforts were made to investigate the cause, and find a solution to fix
it. For details of the motivation, see issue #954;
2. To improve the analysis of MESONET observations of Temperature, the
GSD Terrain Match adjustment, which was used for surface obs of T
(kx=181 & 187), is extended to Mesonet obs of T (kx=188). For details.
see issue #955.